+++ ext_emconf.php
@@ -13,7 +13,7 @@ $EM_CONF[$_EXTKEY] = [
     'constraints' => [
         'depends' => [
             'schema' => '1.7.0-2.99.99',
-            'typo3' => '9.5.16-11.5.99',
+            'typo3' => '10.4.11-11.5.99',
         ],
         'conflicts' => [],
         'suggests' => [],