+++ ext_localconf.php
@@ -1,18 +1,4 @@
 <?php
-defined('TYPO3_MODE') or die();
+defined('TYPO3') or die();
 
-(function() {
-    if ((new TYPO3\CMS\Core\Information\Typo3Version())->getMajorVersion() === 9) {
-        $signalSlotDispatcher = TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(
-            TYPO3\CMS\Extbase\SignalSlot\Dispatcher::class
-        );
-        $signalSlotDispatcher->connect(
-            Brotkrueml\Schema\Core\Model\AbstractType::class,
-            'registerAdditionalTypeProperties',
-            Brotkrueml\SchemaPending\EventListener\RegisterAdditionalProperties::class,
-            '__invoke'
-        );
-    }
-
-    $GLOBALS['TYPO3_CONF_VARS']['SYS']['fluid']['namespaces']['schema'][] = 'Brotkrueml\\SchemaPending\\ViewHelpers';
-})();
+$GLOBALS['TYPO3_CONF_VARS']['SYS']['fluid']['namespaces']['schema'][] = 'Brotkrueml\\SchemaPending\\ViewHelpers';